add tests for pingif state

+++ codenamesbot/state.py
@@ -2,6 +2,7 @@ import collections
 import enum
 import math
 import random
+from typing import Dict, List
 
 from .words import WORDS
 
@@ -470,51 +471,63 @@ class Game:
 class Pingifs:
 
     # just remove all this file shit later.
-    from os.path import join, dirname, parent
+    from os.path import join, dirname
     save_file = join(dirname(dirname(__file__)), 'pingifs.txt')
 
-    def __init__(self):
-        self.__player_map = {}
-        self.__ping_count_map = collections.defaultdict(list)
+    def __init__(self, persistent: bool = True):
+        self.__player_map: Dict[str, int] = {}
+        self.__ping_count_map: Dict[int, str] = collections.defaultdict(list)
 
-        self.__load_persistent_state()
+        self.__persistent: bool = persistent
 
-    def get_players_with_ping_count(ping_count):
+        if self.__persistent:
+            self.__load_persistent_state()
+
+    def get_player_names_with_ping_count(self, ping_count: int) -> List[str]:
         return self.__ping_count_map[ping_count]
 
-    def __getitem__(self, player):
+    def __getitem__(self, player: Player) -> int:
         return self.__player_map[player.name]
 
-    def __setitem__(self, player, ping_count):
+    def __setitem__(self, player: Player, ping_count: int) -> None:
         self.__player_map[player.name] = ping_count
         self.__ping_count_map[ping_count].append(player.name)
-        self.__save_player_count(player.name, ping_count)
 
-    def __contains__(self, player):
+        if self.__persistent:
+            self.__save_player_count(player.name, ping_count)
+
+    def __contains__(self, player: Player) -> bool:
         return player.name in self.__player_map
 
-    def __del__(self, player):
+    def __delitem__(self, player: Player) -> None:
         # Can KeyError, let it be handled by the caller.
         ping_count = self.__player_map[player.name]
 
         del self.__player_map[player.name]
         self.__ping_count_map[ping_count].remove(player.name)
-        self.__delete_player_ping_count(player.name)
 
-    def __load_persistent_state(self):
-        with open(self.save_file, 'r') as f:
-            for line in f:
-                player_name, ping_count = f.split(' ', 1)
+        if self.__persistent:
+            self.__delete_player_ping_count(player.name)
 
-                try:
-                    ping_count = int(ping_count)
-                except ValueError:
-                    raise ValueError("Corrupted pingifs file; could not parse ping count.")
+    def __load_persistent_state(self) -> None:
+        try:
+            with open(self.save_file, 'r') as f:
+                saved_lines = [l.rstrip() for l in f.readlines()]
+        except FileNotFoundError:
+            return
 
-                self.__player_map[player_name] = ping_count
-                self.__ping_count_map[ping_count].append(player_name)
+        for line in saved_lines:
+            player_name, ping_count = f.split(' ', 1)
 
-    def __save_player_count(self, _player_name, _ping_count):
+            try:
+                ping_count = int(ping_count)
+            except ValueError:
+                raise ValueError("Corrupted pingifs file; could not parse ping count.")
+
+            self.__player_map[player_name] = ping_count
+            self.__ping_count_map[ping_count].append(player_name)
+
+    def __save_player_count(self, _player_name: str, _ping_count: int) -> None:
         # just write dict to file rn, completely ignore args.
         # when there is a db, this method will make more sense.
 
@@ -524,7 +537,7 @@ class Pingifs:
             for player_name, ping_count in self.__player_map.items():
                 f.write(f"{player_name} {ping_count}\n")
 
-    def __delete_player_ping_count(self, _player_name):
+    def __delete_player_ping_count(self, _player_name: str):
         # don't actually do anything special yet.
         # when there is a db, this method will make more sense.
